requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Review Scraping
google-play-scraper==1.2.1
app-store-scraper==0.3.5
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
feedparser==6.0.10

# Text Processing
regex==2023.10.3
emoji==2.8.0

# Scheduling
apscheduler==3.10.4
celery==5.3.4
redis==5.0.1

# LLM - Gemini
google-generativeai>=0.3.0

# Email
sendgrid==6.11.0
jinja2==3.1.2

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

